client/src/main/java/io/steem/client/model/operation/AccountWitnessVoteOperation.java
package io.steem.client.model.operation;

import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.databind.PropertyNamingStrategy;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import io.steem.client.model.SteemOperation;
import io.steem.client.model.util.ObjectMapUtils;
import lombok.Builder;

import java.util.Map;

public class AccountWitnessVoteOperation extends SteemOperation {

  @Builder
  @lombok.Value
  @JsonInclude(JsonInclude.Include.NON_NULL)
  @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
  public static class Value {
    private final String account;
    private final String witness;
    private final boolean approve;
  }

  private final Value value;

  private AccountWitnessVoteOperation(Value value) {
    super("account_witness_vote");
    this.value = value;
  }

  public static AccountWitnessVoteOperation of(Value value) {
    return new AccountWitnessVoteOperation(value);
  }

  @Override
  protected Map<String, Object> getValueMap() {
    return ObjectMapUtils.toObjectMap(value);
  }
}
